Move placeBlock to its own file
