+++ lib/plugins/inventory.js
@@ -307,66 +307,6 @@ function inject (bot, { version, hideErrors }) {
     return window
   }
 
-  async function placeBlock (referenceBlock, faceVector) {
-    if (!bot.heldItem) throw new Error('must be holding an item to place a block')
-    // Look at the center of the face
-    const dx = 0.5 + faceVector.x * 0.5
-    const dy = 0.5 + faceVector.y * 0.5
-    const dz = 0.5 + faceVector.z * 0.5
-    await bot.lookAt(referenceBlock.position.offset(dx, dy, dz), false)
-    // TODO: tell the server that we are sneaking while doing this
-    bot.swingArm()
-    const pos = referenceBlock.position
-
-    if (bot.supportFeature('blockPlaceHasHeldItem')) {
-      bot._client.write('block_place', {
-        location: pos,
-        direction: vectorToDirection(faceVector),
-        heldItem: Item.toNotch(bot.heldItem),
-        cursorX: 8,
-        cursorY: 8,
-        cursorZ: 8
-      })
-    } else if (bot.supportFeature('blockPlaceHasHandAndIntCursor')) {
-      bot._client.write('block_place', {
-        location: pos,
-        direction: vectorToDirection(faceVector),
-        hand: 0,
-        cursorX: 8,
-        cursorY: 8,
-        cursorZ: 8
-      })
-    } else if (bot.supportFeature('blockPlaceHasHandAndFloatCursor')) {
-      bot._client.write('block_place', {
-        location: pos,
-        direction: vectorToDirection(faceVector),
-        hand: 0,
-        cursorX: 0.5,
-        cursorY: 0.5,
-        cursorZ: 0.5
-      })
-    } else if (bot.supportFeature('blockPlaceHasInsideBlock')) {
-      bot._client.write('block_place', {
-        location: pos,
-        direction: vectorToDirection(faceVector),
-        hand: 0,
-        cursorX: 0.5,
-        cursorY: 0.5,
-        cursorZ: 0.5,
-        insideBlock: false
-      })
-    }
-
-    const dest = pos.plus(faceVector)
-    const eventName = `blockUpdate:${dest}`
-
-    const [oldBlock, newBlock] = await once(bot, eventName)
-
-    if (oldBlock.type === newBlock.type) {
-      throw new Error(`No block has been placed : the block is still ${oldBlock.name}`)
-    }
-  }
-
   function createActionNumber () {
     nextActionNumber = nextActionNumber === 32767 ? 1 : nextActionNumber + 1
     return nextActionNumber
@@ -569,7 +509,6 @@ function inject (bot, { version, hideErrors }) {
   bot.activateBlock = callbackify(activateBlock)
   bot.activateEntity = callbackify(activateEntity)
   bot.activateEntityAt = callbackify(activateEntityAt)
-  bot.placeBlock = callbackify(placeBlock)
   bot.consume = callbackify(consume)
   bot.activateItem = activateItem
   bot.deactivateItem = deactivateItem
@@ -584,21 +523,4 @@ function inject (bot, { version, hideErrors }) {
   bot.openEntity = callbackify(openEntity)
   bot.moveSlotItem = callbackify(moveSlotItem)
   bot.updateHeldItem = updateHeldItem
-}
-
-function vectorToDirection (v) {
-  if (v.y < 0) {
-    return 0
-  } else if (v.y > 0) {
-    return 1
-  } else if (v.z < 0) {
-    return 2
-  } else if (v.z > 0) {
-    return 3
-  } else if (v.x < 0) {
-    return 4
-  } else if (v.x > 0) {
-    return 5
-  }
-  assert.ok(false, `invalid direction vector ${v}`)
 }
